Disable H on numpad layer
Trialling mod+numpad for window management, as I don't expect this to
conflict with any standard shortcuts, and will be usable on regular
104-key keyboards.

On this Moonlander layout, muscle memory has me trying to use vim-style
movement keys. In a later commit, I will copy the numpad keys to the
window management layer in the layout I expect.