Attempt to improve oled layer status readability
The font on the oled displays is very small, and can't be easily
increased at the moment. When the following PR merges, I'll take another
look at this. https://github.com/qmk/qmk_firmware/pull/19827

Until then, this change removes some duplication/boilerplate, and makes
it a little easier to differentiate the layers at a glance.